Make survey executable from graphile-worker
